Remove `private` declaration as a work-around
When building with Typescript 4 we get the following build error for all private properties and functions:

```
oled.ts - error TS4094: Property '[all private variables and functions]' of exported class expression may not be private or protected.
```

If I read this issue correctly, the `private` definition is not really supported, so I'm commenting it out for now.