Expose `updateDirty` as a public API
If we want to optimize drawing calls, we can use the sync/immediate flag to turn immediate updates off and then only update changes bytes in the last drawing call. However, instead of always checking the flag on the last call, we could also use the updateDirty method as the last function call which would make fiddling with other draw calls a little easier.